Use overpass_query_with_retry() in street_housenumbers_update_result_json()
Instead of a copy&paste loop that naively sleeps for 1 second between
retry requests insead of asking the overpass API to find out the amount
of needed sleep.

This uncovered that overpass_query_with_retry() had a bug: if we really
fail 20 times in a row, then the return value was not an error, but a
success with an empty string. This is unlikely in practice, so change
this to return an actual error, which allows simplifying
update_stats_overpass() and update_settlement_stats_overpass().
