+++ src/overpass_query.rs
@@ -12,6 +12,12 @@
 
 use crate::context;
 
+#[cfg(not(test))]
+use log::info;
+
+#[cfg(test)]
+use std::println as info;
+
 /// Posts the query string to the overpass API and returns the result string.
 pub fn overpass_query(ctx: &context::Context, query: &str) -> anyhow::Result<String> {
     let url = ctx.get_ini().get_overpass_uri() + "/api/interpreter";
@@ -52,6 +58,46 @@ pub fn overpass_query_need_sleep(ctx: &context::Context) -> i32 {
         return 0;
     }
     sleep
+}
+
+/// Sleeps to respect overpass rate limit.
+pub fn overpass_sleep(ctx: &context::Context) {
+    loop {
+        let sleep = overpass_query_need_sleep(ctx);
+        if sleep == 0 {
+            break;
+        }
+        info!("overpass_sleep: waiting for {sleep} seconds");
+        ctx.get_time().sleep(sleep as u64);
+    }
+}
+
+/// Decides if we should retry a query or not.
+pub fn should_retry(retry: i32) -> bool {
+    retry < 20
+}
+
+pub fn overpass_query_with_retry(ctx: &context::Context, query: &str) -> anyhow::Result<String> {
+    let mut retry = 0;
+    let mut ret: anyhow::Result<String> = Ok("".to_string());
+    while should_retry(retry) {
+        if retry > 0 {
+            info!("overpass_query_with_retry: try #{retry}");
+        }
+        retry += 1;
+        overpass_sleep(ctx);
+        let response = match overpass_query(ctx, query) {
+            Ok(value) => value,
+            Err(err) => {
+                info!("overpass_query_with_retry: http error: {err}");
+                ret = Err(err);
+                continue;
+            }
+        };
+
+        return Ok(response);
+    }
+    ret
 }
 
 #[cfg(test)]
